requirements.txt
Django[argon2]==2.0.6
django-enumfields==0.10.1
django-environ==0.4.4
djangorestframework==3.8.2

python-twitter==3.4.2
